[Feature][SIG-27426] Add hooks to fetch metadata for Databricks queries (#14)
To make debugging Databricks queries easier, add some basic hooks to
fetch metadata for queries (e.g. session ID, query ID) from Databricks
queries. The intention here is that Multiplex will register callbacks
using `WithOpenSessionHook` and `WithOperationMetadataHook` that will
store the metadata that Databricks returns as trace tags.

I went with this approach instead of stuffing the metadata in the
Databricks driver's `Rows` struct, mainly because `Rows` doesn't always
get returned, e.g. when executing statements that don't return any data,
or when there are errors.

While the goal here is to simplify debugging failed queries, it's
possible for a Databricks query to fail without giving us a query ID,
namely, if `ExecuteStatement` returns an error. This happens when the
query contains a syntax error, for example. In that case, we'll at least
have the session ID, but we can continue to iterate on fetching query
IDs when `ExecuteStatement` fails. However, this will enable us to get
query IDs for queries that fail at any later point.

Signed-off-by: Eric Bannatyne <eric@sigmacomputing.com>